split Service Remoting integration into client and service
